+++ AssetTracker/Models/HistoricalData.cs
@@ -5,10 +5,10 @@ namespace AssetTracker.Models
 	{
         
         public DateTime Date { get; set; }
-        public double ClosePrice { get; set; }
-        public double Low { get; set; }
-        public double High { get; set; }
-        public double Open { get; set; }
+        public decimal ClosePrice { get; set; }
+        public decimal Low { get; set; }
+        public decimal High { get; set; }
+        public decimal Open { get; set; }
         public double Volume { get; set; }
         
         public HistoricalData()
